src/train/train.py
import os
import yaml
import torch
from transformers import AutoTokenizer, AutoModelForSequenceClassification, Trainer, TrainingArguments
# Placeholder imports for PEFT modules
# from src.models.peft_controller import MetaPeftController

def load_config(path="src/config/default.yaml"):
    with open(path, "r") as f:
        return yaml.safe_load(f)

def main():
    # Load configuration
    cfg = load_config()
    os.makedirs(cfg["output_dir"], exist_ok=True)
    
    # Set seed and device
    torch.manual_seed(cfg["seed"])
    device = torch.device(cfg["device"] if torch.cuda.is_available() else "cpu")
    
    # Tokenizer & Model
    tokenizer = AutoTokenizer.from_pretrained(cfg["model_name"])
    model = AutoModelForSequenceClassification.from_pretrained(cfg["model_name"], num_labels=2)
    model.to(device)
    
    # TODO: Initialize MetaPeftController with cfg axes
    
    # Data preparation (example using GLUE MRPC)
    from datasets import load_dataset, load_metric
    dataset = load_dataset("glue", cfg["task_name"])
    metric = load_metric("glue", cfg["task_name"])
    
    def preprocess(examples):
        return tokenizer(examples["sentence1"], examples["sentence2"], truncation=True, padding="max_length",
                         max_length=cfg["max_seq_length"])
    
    tokenized = dataset.map(preprocess, batched=True)
    tokenized = tokenized.rename_column("label", "labels")
    tokenized.set_format("torch", columns=["input_ids","attention_mask","labels"])
    
    # Training arguments
    training_args = TrainingArguments(
        output_dir=cfg["output_dir"],
        evaluation_strategy="steps",
        eval_steps=cfg["logging_steps"],
        logging_steps=cfg["logging_steps"],
        per_device_train_batch_size=cfg["train_batch_size"],
        per_device_eval_batch_size=cfg["eval_batch_size"],
        learning_rate=cfg["learning_rate"],
        num_train_epochs=cfg["num_train_epochs"],
        seed=cfg["seed"],
        save_steps=cfg["save_steps"],
    )
    
    trainer = Trainer(
        model=model,
        args=training_args,
        train_dataset=tokenized["train"],
        eval_dataset=tokenized["validation"],
        tokenizer=tokenizer,
        # callbacks or custom loop for PEFT controller integration
    )
    
    # Start training
    trainer.train()

if __name__ == "__main__":
    main()
